Server/requirement.txt
psycog2

django

djangorestframework

djangorestframework-simplejwt

django-cors-headers